# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Server API
  version: 0.0.1
servers:
  - url: "http://localhost:13337/"
    description: "Development"
components:
  securitySchemes:
    serverAuth:
      type: http
      scheme: bearer
  responses:
    ResponseError:
      description: Error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
    ServerBase:
      type: object
      properties:
        name:
          type: string
        count:
          type: integer
        port:
          type: integer
    ServerAuth:
      type: object
      required:
        - password
      properties:
        password:
          type: string
    ServerInput:
      allOf:
        - $ref: "#/components/schemas/ServerBase"
        - type: object
          required:
            - name
            - port
            - password
          properties:
            password:
              type: string
    ServerOutput:
      allOf:
        - $ref: "#/components/schemas/ServerBase"
        - type: object
          required:
            - id
            - name
            - host
            - port
            - count
          properties:
            id:
              type: string
            host:
              type: string
            time:
              type: integer
    ServerOutputToken:
      allOf:
        - $ref: "#/components/schemas/ServerOutput"
        - type: object
          required:
            - token
          properties:
            token:
              type: string
paths:
  /ping:
    get:
      description: Checks if server is active.
      responses:
        "200":
          description: A pong response.
          content:
            application/json:
              schema:
                type: boolean
        default:
          $ref: "#/components/responses/ResponseError"
  /servers/:
    get:
      description: Returns all active Servers.
      responses:
        "200":
          description: A list of Servers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServerOutput"
        default:
          $ref: "#/components/responses/ResponseError"
    post:
      description: Creates a new Server.
      requestBody:
        required: true
        content:
          application/json:
            schema:
               $ref: "#/components/schemas/ServerInput"
      responses:
        "200":
          description: A new Server + Auth.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerOutputToken"
        default:
          $ref: "#/components/responses/ResponseError"
  /servers/{id}:
    get:
      description: Returns basic server details.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Server details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerOutput"
        default:
          $ref: "#/components/responses/ResponseError"
    post:
      description: Returns extended server details.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerAuth"
      responses:
        "200":
          description: Server details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerOutput"
        default:
          $ref: "#/components/responses/ResponseError"
    put:
      description: Updates server details.
      security:
        - serverAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
               $ref: "#/components/schemas/ServerBase"
      responses:
        "200":
          description: Server details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerOutput"
        default:
          $ref: "#/components/responses/ResponseError"
    delete:
      description: Deletes server.
      security:
        - serverAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Server details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerOutput"
        default:
          $ref: "#/components/responses/ResponseError"